+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 This is the official BubbleOS changelog! All features will be recorded in _'groups'_, and the latest version of that _'group'_ will have an executable and a release paired with them.
 
+## Build 198 to 200 (v1.9.8-beta to v2.0.0)
+
+### Added Features
+
+- Added the ability to silence most success messages (customizable using `setmgr`).
+
+### Changed/Fixed Features
+
+- Changed the timeout in the `ping` command from 15 seconds to 10 seconds.
+- Fixed an issue where pressing Ctrl+C while having exit prompts enabled would cause BubbleOS to freeze if "no" was entered.
+- Fixed an issue where pressing Ctrl+C while on the prompt in `mkfile` would crash BubbleOS.
+- Fixed an issue where pressing Ctrl+C while a command was executing would not exit BubbleOS, and freeze the shell instead.
+
+### Removed Features
+
+- Removed the silence (usually `-s`) flags from most commands in favor of the setting in `setmgr`.
+
 ## Build 180 to 198 (v1.8.0-beta to v1.9.8-beta)
 
 ### Added Features